requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.13
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
factory-boy==3.2.1
Faker==13.4.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
whitenoise==5.3.0
